refactor: separate to handle response
